SurveiLens-auth0-fixed/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
pyjwt[crypto]==2.9.0
python-jose==3.3.0
requests==2.32.3
redis==5.0.7
celery[redis]==5.4.0
snowflake-connector-python==3.11.1
boto3==1.34.162
ultralytics==8.3.24
opencv-python-headless==4.10.0.84
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.2.post1
webrtcvad==2.0.10
vosk==0.3.45
simplejson==3.19.3
pydantic==2.8.2
solana==0.30.2
base58==2.1.1
elevenlabs==1.6.1
eventlet==0.36.1
sse-starlette==2.1.0
